+++ include/scaluq/gate/gate_pauli.hpp
@@ -2,7 +2,6 @@
 
 #include <vector>
 
-#include "../operator/apply_pauli.hpp"
 #include "../operator/pauli_operator.hpp"
 #include "../util/utility.hpp"
 #include "gate.hpp"
@@ -27,23 +26,9 @@ public:
     }
     internal::ComplexMatrix<Fp> get_matrix() const override { return this->_pauli.get_matrix(); }
 
-    void update_quantum_state(StateVector<Fp>& state_vector) const override {
-        auto [bit_flip_mask, phase_flip_mask] = _pauli.get_XZ_mask_representation();
-        apply_pauli(
-            this->_control_mask, bit_flip_mask, phase_flip_mask, _pauli.coef(), state_vector);
-    }
+    void update_quantum_state(StateVector<Fp>& state_vector) const override;
 
-    std::string to_string(const std::string& indent) const override {
-        std::ostringstream ss;
-        auto controls = this->control_qubit_list();
-        ss << indent << "Gate Type: Pauli\n";
-        ss << indent << "  Control Qubits: {";
-        for (std::uint32_t i = 0; i < controls.size(); ++i)
-            ss << controls[i] << (i == controls.size() - 1 ? "" : ", ");
-        ss << "}\n";
-        ss << indent << "  Pauli Operator: \"" << _pauli.get_pauli_string() << "\"";
-        return ss.str();
-    }
+    std::string to_string(const std::string& indent) const override;
 };
 
 template <std::floating_point Fp>
@@ -66,37 +51,11 @@ public:
             this->_control_mask, _pauli, -_angle);
     }
 
-    internal::ComplexMatrix<Fp> get_matrix() const override {
-        internal::ComplexMatrix<Fp> mat = this->_pauli.get_matrix_ignoring_coef();
-        Complex<Fp> true_angle = _angle * _pauli.coef();
-        StdComplex<Fp> imag_unit(0, 1);
-        mat = (StdComplex<Fp>)Kokkos::cos(-true_angle / 2) *
-                  internal::ComplexMatrix<Fp>::Identity(mat.rows(), mat.cols()) +
-              imag_unit * (StdComplex<Fp>)Kokkos::sin(-true_angle / 2) * mat;
-        return mat;
-    }
-    void update_quantum_state(StateVector<Fp>& state_vector) const override {
-        auto [bit_flip_mask, phase_flip_mask] = _pauli.get_XZ_mask_representation();
-        apply_pauli_rotation(this->_control_mask,
-                             bit_flip_mask,
-                             phase_flip_mask,
-                             _pauli.coef(),
-                             _angle,
-                             state_vector);
-    }
+    internal::ComplexMatrix<Fp> get_matrix() const override;
 
-    std::string to_string(const std::string& indent) const override {
-        std::ostringstream ss;
-        auto controls = this->control_qubit_list();
-        ss << indent << "Gate Type: PauliRotation\n";
-        ss << indent << "  Angle: " << _angle << "\n";
-        ss << indent << "  Control Qubits: {";
-        for (std::uint32_t i = 0; i < controls.size(); ++i)
-            ss << controls[i] << (i == controls.size() - 1 ? "" : ", ");
-        ss << "}\n";
-        ss << indent << "  Pauli Operator: \"" << _pauli.get_pauli_string() << "\"";
-        return ss.str();
-    }
+    void update_quantum_state(StateVector<Fp>& state_vector) const override;
+
+    std::string to_string(const std::string& indent) const override;
 };
 }  // namespace internal
