--- include/scaluq/gate/update_ops_dense_matrix.cpp
@@ -1,10 +1,0 @@
-#include <Kokkos_Core.hpp>
-#include <Kokkos_StdAlgorithms.hpp>
-
-#include "../types.hpp"
-#include "../util/utility.hpp"
-#include "update_ops.hpp"
-
-namespace scaluq {
-namespace internal {}  // namespace internal
-}  // namespace scaluq
