--- include/scaluq/gate/update_ops_standard.cpp
@@ -1,7 +1,0 @@
-#include <Kokkos_Core.hpp>
-#include <Kokkos_StdAlgorithms.hpp>
-
-#include "../types.hpp"
-#include "../util/utility.hpp"
-#include "constant.hpp"
-#include "update_ops.hpp"
