--- include/scaluq/operator/apply_pauli.cpp
@@ -1,9 +1,0 @@
-#include "apply_pauli.hpp"
-
-#include <Kokkos_Core.hpp>
-
-#include "../constant.hpp"
-#include "../types.hpp"
-#include "../util/utility.hpp"
-
-namespace scaluq::internal {}  // namespace scaluq::internal
