--- include/scaluq/operator/operator.cpp
@@ -1,8 +1,0 @@
-#include "./operator.hpp"
-
-#include <ranges>
-
-#include "../constant.hpp"
-#include "../util/utility.hpp"
-
-namespace scaluq {}  // namespace scaluq
