--- include/scaluq/operator/pauli_operator.cpp
@@ -1,6 +1,0 @@
-#include "pauli_operator.hpp"
-
-#include "../constant.hpp"
-#include "../util/utility.hpp"
-
-namespace scaluq {}  // namespace scaluq
