--- include/scaluq/state/state_vector.cpp
@@ -1,3 +1,0 @@
-#include "state_vector.hpp"
-
-namespace scaluq {}  // namespace scaluq
