--- include/scaluq/state/state_vector_batched.cpp
@@ -1,1 +1,0 @@
-#include "state_vector_batched.hpp"
